Added event for hall door closing scenario
