+++ src/main/java/ru/sbt/mipt/oop/Application.java
@@ -2,6 +2,7 @@ package ru.sbt.mipt.oop;
 
 import ru.sbt.mipt.oop.events.EventProcessor;
 import ru.sbt.mipt.oop.events.SensorEvent;
+import ru.sbt.mipt.oop.events.handlers.CloseHallDoorEventHandler;
 import ru.sbt.mipt.oop.events.handlers.DoorEventHandler;
 import ru.sbt.mipt.oop.events.handlers.LightEventHandler;
 import ru.sbt.mipt.oop.events.producers.EventProducer;
@@ -9,18 +10,18 @@ import ru.sbt.mipt.oop.events.producers.RandomEventProducer;
 import ru.sbt.mipt.oop.serialization.SmartHomeDeserializer;
 import ru.sbt.mipt.oop.serialization.SmartHomeJsonDeserializer;
 
-import java.io.IOException;
 import java.util.Arrays;
 
 public class Application {
 
-    public static void main(String... args) throws IOException {
+    public static void main(String... args) {
         SmartHomeDeserializer deserializer = new SmartHomeJsonDeserializer("output.js");
         SmartHome smartHome = deserializer.deserialize();
 
         EventProcessor eventProcessor = new EventProcessor(Arrays.asList(
                 new DoorEventHandler(smartHome),
-                new LightEventHandler(smartHome)
+                new LightEventHandler(smartHome),
+                new CloseHallDoorEventHandler(smartHome)
         ));
 
         EventProducer eventProducer = new RandomEventProducer();
